refactor: move run_enumerator into example_datasets service
